rewrite inline styles to mui classes; move mock data from components to utils
