+++ packages/client/src/components/Forum/TopicItem.tsx
@@ -4,6 +4,7 @@ import { Card, CardContent, Typography, Box, Avatar } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
 
 import { Comment as CommentIcon } from '@material-ui/icons'
+import { Topic } from '../../types/types'
 const useStyles = makeStyles(theme => ({
   card: {
     marginBottom: theme.spacing(2),
@@ -31,16 +32,11 @@ const useStyles = makeStyles(theme => ({
     textDecoration: 'none',
     color: 'inherit',
   },
+  commentIcon: {
+    marginRight: 4,
+    marginLeft: 8,
+  },
 }))
-
-export interface Topic {
-  id: number
-  title: string
-  content: string
-  author: string
-  createdAt: Date
-  commentsCount: number
-}
 
 interface TopicItemProps {
   topic: Topic
@@ -84,10 +80,7 @@ export const TopicItem: React.FC<TopicItemProps> = ({
             <Typography variant="caption" color="textSecondary">
               {topic.createdAt.toLocaleDateString()}
             </Typography>
-            <CommentIcon
-              fontSize="small"
-              style={{ marginRight: 4, marginLeft: 8 }}
-            />
+            <CommentIcon fontSize="small" className={classes.commentIcon} />
             <Typography variant="caption" color="textSecondary">
               {topic.commentsCount}
             </Typography>
@@ -97,42 +90,3 @@ export const TopicItem: React.FC<TopicItemProps> = ({
     </Card>
   )
 }
-
-export const mockTopics: Topic[] = [
-  {
-    id: 1,
-    title: 'Тактика за Бегуна (Зеленый) vs Тактика за Снайпера (Красный)',
-    content:
-      'Мы с другом постоянно спорим, за кого эффективнее играть. Я говорю, что скорость Зеленого бесценна для разведки. Он утверждает, что способность Красного убирать зомби с дальней дистанции спасает команду. Что вы думаете? Как вам кажется, чья способность чаще приносит победу?',
-    author: 'ryl73',
-    createdAt: new Date(2024, 0, 15),
-    commentsCount: 4,
-  },
-  {
-    id: 2,
-    title: 'Паника! На кубике выпала 4. Все пропало?',
-    content:
-      'Только что проиграли партию из-за злосчастной четверки на кубике в самый неподходящий момент. Активировались ВСЕ зомби на поле, и нас просто сомнули у самого выхода. Как вы минимизируете риски выпадения четверки? Может, есть карты, которые могут это как-то заблокировать?',
-    author: 'andreissh',
-    createdAt: new Date(2024, 0, 16),
-    commentsCount: 3,
-  },
-  {
-    id: 3,
-    title: 'Карта "Сгоряча" — самая переоцененная или самая полезная?',
-    content:
-      'Вчера был жаркий спор. Я считаю, что карта "Сгоряча" (два движения за один ход) — это must-have для доставки тотема к выходу в финале. Друг говорит, что она бесполезна и лучше взять любое оружие. Кто прав? Как вы используете эту карту?',
-    author: 'cyperus-papyrus',
-    createdAt: new Date(2024, 0, 17),
-    commentsCount: 5,
-  },
-  {
-    id: 4,
-    title: 'Раскидаться или сгруппироваться? Стратегия первых ходов.',
-    content:
-      'После стартового размещения зомби мы всегда стоим перед выбором: бежать всем в одну сторону или раскидаться по углам для быстрого поиска выхода и тотема? В первом случае мы медленные, но безопасные. Во втором — нас быстро изолируют и выводят поодиночке. Какой ваш проверенный план?',
-    author: 'MarsiKris76',
-    createdAt: new Date(2024, 0, 18),
-    commentsCount: 2,
-  },
-]
