+++ packages/client/src/components/Game/StartDialog.tsx
@@ -20,12 +20,12 @@ type Props = {
 
 const useStyles = makeStyles(() => ({
   root: {
-    '& .MuiOutlinedInput-root': {
-      width: '100%',
-    },
     '& .MuiButton-root': {
       margin: '0',
     },
+  },
+  columnHeader: {
+    textAlign: 'center',
   },
 }))
 
@@ -50,8 +50,8 @@ export const StartDialog: FC<Props> = ({ isDialog, startGame }) => {
         </Box>
 
         <Box display="grid" gridTemplateColumns="1fr 1px 1fr" gridGap="24px">
-          <div className="">
-            <Typography variant="h6" style={{ textAlign: 'center' }}>
+          <div>
+            <Typography variant="h6" className={classes.columnHeader}>
               Играть по сети
             </Typography>
             <Box
@@ -87,7 +87,7 @@ export const StartDialog: FC<Props> = ({ isDialog, startGame }) => {
           </div>
           <Divider orientation="vertical" flexItem />
           <div className="">
-            <Typography variant="h6" style={{ textAlign: 'center' }}>
+            <Typography variant="h6" className={classes.columnHeader}>
               Играть локально
             </Typography>
             <Box
