+++ packages/client/src/pages/MainPage.tsx
@@ -12,7 +12,6 @@ import {
   Grid,
   Card,
   CardContent,
-  useTheme,
 } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
 import { logout } from '../api/LoginAPI'
@@ -21,6 +20,39 @@ import { useAuth } from '../hooks/useAuth'
 import { FullscreenToggle } from '../components/FullscreenToggle/FullscreenToggle'
 
 const useStyles = makeStyles(theme => ({
+  wrapper: {
+    backgroundColor: theme.palette.background.default,
+    paddingBottom: theme.spacing(5),
+  },
+  startGameBtn: {
+    width: '100%',
+    maxWidth: '435px',
+    fontWeight: 'bold',
+    textTransform: 'uppercase',
+    letterSpacing: 1,
+  },
+  card: {
+    height: '100%',
+  },
+  cardContent: {
+    padding: 24,
+  },
+  featureTitle: {
+    whiteSpace: 'nowrap',
+    overflow: 'hidden',
+    textOverflow: 'ellipsis',
+  },
+  featureDescription: {
+    color: theme.palette.text.secondary,
+    marginTop: '20px',
+    fontSize: '18px',
+  },
+  readyBlockTitle: {
+    marginBottom: '2rem',
+    marginTop: '4rem',
+    fontWeight: 'bold',
+    textAlign: 'center',
+  },
   firstScreen: {
     minHeight: 500,
     backgroundImage: 'url(/src/assets/landing-first-screen.webp)',
@@ -56,7 +88,7 @@ const useStyles = makeStyles(theme => ({
       fontSize: '2rem',
     },
   },
-  readyButton: {
+  readyBlockButton: {
     maxWidth: '250px',
     width: '100%',
     padding: '12px 24px',
@@ -64,11 +96,31 @@ const useStyles = makeStyles(theme => ({
   },
 }))
 
+const features = [
+  {
+    icon: <span>🔍</span>,
+    title: 'Исследуй',
+    description:
+      'Обыскивай комнаты в\u00A0поисках оружия, аптечек и\u00A0ключей к\u00A0выходу. Каждый шаг может быть последним',
+  },
+  {
+    icon: <span>⚔️</span>,
+    title: 'Сражайся',
+    description:
+      'Отбивайся от\u00A0полчищ мертвецов. Используй всё, что\u00A0попадётся под\u00A0руку\u00A0— от\u00A0биты до\u00A0дробовика',
+  },
+  {
+    icon: <span>🛡️</span>,
+    title: 'Выживай',
+    description:
+      'Держись вместе с\u00A0другими выжившими. Только команда поможет вам\u00A0дожить до\u00A0рассвета',
+  },
+]
+
 export const MainPage = () => {
   usePage({ initPage: initMainPage })
   const { isLoggedIn, clearUser } = useAuth()
   const navigate = useNavigate()
-  const theme = useTheme()
   const classes = useStyles()
 
   useEffect(() => {
@@ -77,29 +129,8 @@ export const MainPage = () => {
     }
   }, [isLoggedIn, navigate])
 
-  const features = [
-    {
-      icon: <span>🔍</span>,
-      title: 'Исследуй',
-      description:
-        'Обыскивай комнаты в\u00A0поисках оружия, аптечек и\u00A0ключей к\u00A0выходу. Каждый шаг может быть последним',
-    },
-    {
-      icon: <span>⚔️</span>,
-      title: 'Сражайся',
-      description:
-        'Отбивайся от\u00A0полчищ мертвецов. Используй всё, что\u00A0попадётся под\u00A0руку\u00A0— от\u00A0биты до\u00A0дробовика',
-    },
-    {
-      icon: <span>🛡️</span>,
-      title: 'Выживай',
-      description:
-        'Держись вместе с\u00A0другими выжившими. Только команда поможет вам\u00A0дожить до\u00A0рассвета',
-    },
-  ]
-
   return (
-    <Box style={{ backgroundColor: theme.palette.background.default }} pb={5}>
+    <Box className={classes.wrapper}>
       <Helmet>
         <meta charSet="utf-8" />
         <title>Главная</title>
@@ -121,13 +152,7 @@ export const MainPage = () => {
             variant="contained"
             size="large"
             color="primary"
-            style={{
-              width: '100%',
-              maxWidth: '435px',
-              fontWeight: 'bold',
-              textTransform: 'uppercase',
-              letterSpacing: 1,
-            }}>
+            className={classes.startGameBtn}>
             Начать игру
           </Button>
         </Container>
@@ -145,29 +170,21 @@ export const MainPage = () => {
           <Grid container spacing={4}>
             {features.map((feature, index) => (
               <Grid item xs={12} md={4} key={index}>
-                <Card style={{ height: '100%' }}>
-                  <CardContent style={{ padding: '24px' }}>
+                <Card className={classes.card}>
+                  <CardContent className={classes.cardContent}>
                     <Box display="flex" alignItems="center" mb={2}>
                       <Box className={classes.iconWrapper}>{feature.icon}</Box>
                       <Typography
                         variant="h6"
                         component="h3"
-                        style={{
-                          whiteSpace: 'nowrap',
-                          overflow: 'hidden',
-                          textOverflow: 'ellipsis',
-                        }}>
+                        className={classes.featureTitle}>
                         {feature.title}
                       </Typography>
                     </Box>
 
                     <Typography
                       variant="body1"
-                      style={{
-                        color: theme.palette.text.secondary,
-                        marginTop: '20px',
-                        fontSize: '18px',
-                      }}>
+                      className={classes.featureDescription}>
                       {feature.description}
                     </Typography>
                   </CardContent>
@@ -180,13 +197,7 @@ export const MainPage = () => {
         <Typography
           variant="h4"
           component="h4"
-          align="center"
-          gutterBottom
-          style={{
-            marginBottom: '2rem',
-            marginTop: '4rem',
-            fontWeight: 'bold',
-          }}>
+          className={classes.readyBlockTitle}>
           Готов проверить себя на прочность?
         </Typography>
 
@@ -196,7 +207,7 @@ export const MainPage = () => {
             to="/game"
             variant="contained"
             color="primary"
-            className={classes.readyButton}>
+            className={classes.readyBlockButton}>
             Я готов!
           </Button>
           <Button
@@ -207,7 +218,7 @@ export const MainPage = () => {
             to="/signin"
             variant="contained"
             color="primary"
-            className={classes.readyButton}>
+            className={classes.readyBlockButton}>
             Выход
           </Button>
         </Box>
