+++ packages/client/src/types/types.ts
@@ -1,0 +1,15 @@
+export interface Topic {
+  id: number
+  title: string
+  content: string
+  author: string
+  createdAt: Date
+  commentsCount: number
+}
+
+export interface Comment {
+  id: number
+  author: string
+  content: string
+  createdAt: Date
+}
